move toaster logic from content script
